v0.95.0
* Added some interpolation algorithms for the zooming into photos, videos, icons and other images. Now the user can choose (in the preferences) from: Area Sampling, Bilinear, Bicubic, Lanczos, and Adaptive.
* Accelerated graphics display by making use of the AVX2 CPU extension.
* Extended the graphics API by a texture class which allows more flexible filling of figures with colors and images.
* Ported the audio/video player to Windows, based on the VLC library.
* Switched from Xine to VLC also for other OSes because some tests showed that VLC is more crash-resistant against randomly corrupted files and thereby most likely more secure.
* Ported the archive viewer to Windows.
* Completely removed support for outdated archive file formats to improve security and simplify maintenance: Goodbye to arc, arj, lha, rar, and zoo.
* Minor fixes and improvements.

+++ src/emCore/emImage.cpp
@@ -1,7 +1,7 @@
 //------------------------------------------------------------------------------
 // emImage.cpp
 //
-// Copyright (C) 2001,2003-2010,2014,2016,2018-2019 Oliver Hamann.
+// Copyright (C) 2001,2003-2010,2014,2016,2018-2020 Oliver Hamann.
 //
 // Homepage: http://eaglemode.sourceforge.net/
 //
@@ -285,7 +285,7 @@ L_Error:
 
 
 void emImage::TryParseTga(
-	const unsigned char * tgaData, int tgaSize, int channelCount
+	const unsigned char * tgaData, size_t tgaSize, int channelCount
 )
 {
 	const unsigned char * tgaEnd, * p;
@@ -1205,7 +1205,7 @@ void emImage::CopyTransformed(
 					switch (icc)
 					{
 					case 1:
-						sp=imap+((ssize_t)sy)*iw+((ssize_t)sx);
+						sp=imap+((int)sy)*(size_t)iw+((int)sx);
 						switch (cc)
 						{
 						case 1:
@@ -1232,7 +1232,7 @@ void emImage::CopyTransformed(
 						}
 						break;
 					case 2:
-						sp=imap+(((ssize_t)sy)*iw+((ssize_t)sx))*2;
+						sp=imap+(((int)sy)*(size_t)iw+((int)sx))*2;
 						switch (cc)
 						{
 						case 1:
@@ -1258,7 +1258,7 @@ void emImage::CopyTransformed(
 						}
 						break;
 					case 3:
-						sp=imap+(((ssize_t)sy)*iw+((ssize_t)sx))*3;
+						sp=imap+(((int)sy)*(size_t)iw+((int)sx))*3;
 						switch (cc)
 						{
 						case 1:
@@ -1285,7 +1285,7 @@ void emImage::CopyTransformed(
 						}
 						break;
 					default:
-						sp=imap+(((ssize_t)sy)*iw+((ssize_t)sx))*4;
+						sp=imap+(((int)sy)*(size_t)iw+((int)sx))*4;
 						switch (cc)
 						{
 						case 1:
